Moved `parse_time()` to `endaq.ide.util`
